Testes de serico de emprestimos terminados

+++ DonaRosangela/DonaRosangela.App/Features/Loans/LoanService.cs
@@ -1,4 +1,5 @@
-﻿using DonaRosangela.Domain.Features.Loans;
+﻿using DonaRosangela.Domain.Exceptions;
+using DonaRosangela.Domain.Features.Loans;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -17,16 +18,19 @@ namespace DonaRosangela.App.Features.Loans
 
         public Loan Add(Loan loan)
         {
+            loan.Validate();
             return _repository.Add(loan);
         }
 
         public void Delete(Loan loan)
         {
+            if (loan.Id == 0) throw new InvalidIdException();
             _repository.Delete(loan.Id);
         }
 
         public Loan Get(Loan loan)
         {
+            if (loan.Id == 0) throw new InvalidIdException();
             return _repository.Get(loan.Id);
         }
 
@@ -37,6 +41,8 @@ namespace DonaRosangela.App.Features.Loans
 
         public Loan Update(Loan loan)
         {
+            if (loan.Id == 0) throw new InvalidIdException();
+            loan.Validate();
             return _repository.Update(loan);
         }
     }
